Clang compat.

+++ tools/hmemorymappedfile.cxx
@@ -55,7 +55,7 @@ HMemoryMappedFile::HMemoryMappedFile( yaal::hcore::HString const& path_, int lon
 		if ( ! size_ ) {
 			M_ENSURE( ( _size = static_cast<int long>( ::lseek( _fd, 0, SEEK_END ) ) ) >= 0 );
 		}
-		M_ENSURE( ( _map = ::mmap( NULL, _size, PROT_READ | PROT_WRITE, MAP_PRIVATE, _fd, 0 ) ) != FWD_MAP_FAILED );
+		M_ENSURE( ( _map = ::mmap( NULL, static_cast<size_t>( _size ), PROT_READ | PROT_WRITE, MAP_PRIVATE, _fd, 0 ) ) != FWD_MAP_FAILED );
 	} catch ( ... ) {
 		M_SAFE( M_ENSURE( ::close( _fd ) != 0 ) );
 		_fd = -1;
@@ -70,7 +70,7 @@ HMemoryMappedFile::~HMemoryMappedFile( void ) {
 	M_PROLOG
 	if ( _fd >= 0 ) {
 		M_ASSERT( _map );
-		M_ENSURE( ::munmap( _map, _size ) == 0 );
+		M_ENSURE( ::munmap( _map, static_cast<size_t>( _size ) ) == 0 );
 		M_ENSURE( ::close( _fd ) == 0 );
 	} else {
 		M_ASSERT( ! _map );
